Fix reviewer feedback
