+++ app.rb
@@ -1,7 +1,8 @@
-require './person'
-require './student'
-require './teacher'
-require './classroom'
+require_relative './person'
+require_relative './student'
+require_relative './teacher'
+require_relative './classroom'
+require_relative './rental'
 
 class App
   def initialize
