+++ classroom.rb
@@ -1,6 +1,9 @@
+require_relative './person'
+require_relative './student'
+require_relative './teacher'
+
 class Classroom
-  attr_accessor :label
-  attr_reader :students
+  attr_accessor :label, :students
 
   def initialize(label)
     @label = label
@@ -10,16 +13,6 @@ class Classroom
   def add_student(student)
     @students << student
     student.classroom = self
-  end
-end
-
-class Student < Person
-  attr_accessor :classroom
-
-  def initialize(name, age, parent_permission: true, classroom: nil)
-    super(name, age, parent_permission)
-    @classroom = classroom
-    classroom&.add_student(self)
   end
 end
 
@@ -35,19 +28,6 @@ class Book
   def add_rental(person, date)
     rental = Rental.new(date, self, person)
     @rentals << rental
-    person.add_rental(rental)
-  end
-end
-
-class Rental
-  attr_accessor :date
-  attr_reader :person, :book
-
-  def initialize(person, book, date)
-    @person = person
-    @book = book
-    @date = date
-    person.rentals << self
-    book.rentals << self
+    rental
   end
 end
