+++ rental.rb
@@ -1,0 +1,11 @@
+class Rental
+  attr_accessor :date, :person, :book
+
+  def initialize(date, person, book)
+    @date = date
+    @person = person
+    @book = book
+    book.rentals << self
+    person.rentals << self
+  end
+end
