+++ student.rb
@@ -1,6 +1,10 @@
+require_relative './person'
+
 class Student < Person
-  def initialize(age, classroom, parent_permission: true, name: 'Unknown')
-    super(age, parent_permission: parent_permission, name: name)
+  attr_accessor :classroom
+
+  def initialize(age, classroom, name, parent_permission: true)
+    super(age, name, parent_permission: parent_permission)
     @classroom = classroom
   end
